+++ examples/multi-module-sample/spring-boot-in-multi-module-sample2/src/main/java/io/sentry/samples/spring/boot/Test.kt
@@ -1,0 +1,3 @@
+package io.sentry.samples.spring.boot
+
+class Test