+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/AndroidComponentsConfig.kt
@@ -28,6 +28,8 @@ import io.sentry.android.gradle.tasks.SentryGenerateProguardUuidTask
 import io.sentry.android.gradle.tasks.SentryUploadProguardMappingsTask
 import io.sentry.android.gradle.tasks.configureNativeSymbolsTask
 import io.sentry.android.gradle.tasks.dependencies.SentryExternalDependenciesReportTaskFactory
+import io.sentry.android.gradle.telemetry.SentryTelemetryService
+import io.sentry.android.gradle.telemetry.withSentryTelemetry
 import io.sentry.android.gradle.transforms.MetaInfStripTransform
 import io.sentry.android.gradle.util.AgpVersions
 import io.sentry.android.gradle.util.AgpVersions.isAGP74
@@ -40,17 +42,18 @@ import io.sentry.android.gradle.util.hookWithMinifyTasks
 import io.sentry.android.gradle.util.info
 import java.io.File
 import org.gradle.api.Project
+import org.gradle.api.provider.Provider
 import org.gradle.api.provider.SetProperty
 import org.gradle.api.tasks.TaskProvider
-import org.gradle.build.event.BuildEventsListenerRegistry
+import org.gradle.internal.build.event.BuildEventListenerRegistryInternal
 
 fun AndroidComponentsExtension<*, *, *>.configure(
     project: Project,
     extension: SentryPluginExtension,
-    listenerRegistry: BuildEventsListenerRegistry,
+    buildEvents: BuildEventListenerRegistryInternal,
     cliExecutable: String,
     sentryOrg: String?,
-    sentryProject: String?
+    sentryProject: String?,
 ) {
     // temp folder for sentry-related stuff
     val tmpDir = File("${project.buildDir}${sep}tmp${sep}sentry")
@@ -60,13 +63,22 @@ fun AndroidComponentsExtension<*, *, *>.configure(
         if (isVariantAllowed(extension, variant.name, variant.flavorName, variant.buildType)) {
             val paths = OutputPaths(project, variant.name)
 
-            variant.configureDependenciesTask(project, extension)
+            val sentryTelemetryProvider = variant.configureTelemetry(
+                project,
+                extension,
+                cliExecutable,
+                sentryOrg,
+                buildEvents
+            )
+
+            variant.configureDependenciesTask(project, extension, sentryTelemetryProvider)
 
             val tasksGeneratingProperties =
                 mutableListOf<TaskProvider<out PropertiesFileOutputTask>>()
             val sourceContextTasks = variant.configureSourceBundleTasks(
                 project,
                 extension,
+                sentryTelemetryProvider,
                 paths,
                 cliExecutable,
                 sentryOrg,
@@ -77,6 +89,7 @@ fun AndroidComponentsExtension<*, *, *>.configure(
             val generateProguardUuidTask = variant.configureProguardMappingsTasks(
                 project,
                 extension,
+                sentryTelemetryProvider,
                 paths,
                 cliExecutable,
                 sentryOrg,
@@ -89,12 +102,18 @@ fun AndroidComponentsExtension<*, *, *>.configure(
             sentryVariant?.configureNativeSymbolsTask(
                 project,
                 extension,
+                sentryTelemetryProvider,
                 cliExecutable,
                 sentryOrg,
                 sentryProject
             )
 
-            variant.configureDebugMetaPropertiesTask(project, tasksGeneratingProperties)
+            variant.configureDebugMetaPropertiesTask(
+                project,
+                extension,
+                sentryTelemetryProvider,
+                tasksGeneratingProperties
+            )
 
             if (extension.tracingInstrumentation.enabled.get()) {
                 /**
@@ -117,7 +136,7 @@ fun AndroidComponentsExtension<*, *, *>.configure(
                  * not be discarded after the configuration phase (where we store the collected
                  * dependencies), and will be passed down to the InstrumentationFactory
                  */
-                listenerRegistry.onTaskCompletion(sentryModulesService)
+                buildEvents.onTaskCompletion(sentryModulesService)
 
                 project.collectModules(
                     "${variant.name}RuntimeClasspath",
@@ -150,6 +169,7 @@ fun AndroidComponentsExtension<*, *, *>.configure(
                 ) {
                     it.sentryModulesService.set(sentryModulesService)
                     it.usesService(sentryModulesService)
+                    it.withSentryTelemetry(extension, sentryTelemetryProvider)
                 }
 
                 variant.artifacts.use(manifestUpdater).wiredWithFiles(
@@ -187,12 +207,16 @@ fun AndroidComponentsExtension<*, *, *>.configure(
 
 private fun Variant.configureDebugMetaPropertiesTask(
     project: Project,
+    extension: SentryPluginExtension,
+    sentryTelemetryProvider: Provider<SentryTelemetryService>,
     tasksGeneratingProperties: List<TaskProvider<out PropertiesFileOutputTask>>
 ) {
     if (isAGP74) {
         val taskSuffix = name.capitalized
         val generateDebugMetaPropertiesTask = SentryGenerateDebugMetaPropertiesTask.register(
             project,
+            extension,
+            sentryTelemetryProvider,
             tasksGeneratingProperties,
             null,
             taskSuffix
@@ -210,9 +234,35 @@ private fun Variant.configureDebugMetaPropertiesTask(
     }
 }
 
+private fun Variant.configureTelemetry(
+    project: Project,
+    extension: SentryPluginExtension,
+    cliExecutable: String,
+    sentryOrg: String?,
+    buildEvents: BuildEventListenerRegistryInternal
+): Provider<SentryTelemetryService> {
+    val variant = if (isAGP74) AndroidVariant74(this) else null
+    val sentryTelemetryProvider = SentryTelemetryService.register(project)
+    project.gradle.taskGraph.whenReady {
+        sentryTelemetryProvider.get().start {
+            SentryTelemetryService.createParameters(
+                project,
+                variant,
+                extension,
+                cliExecutable,
+                sentryOrg,
+                "Android"
+            )
+        }
+        buildEvents.onOperationCompletion(sentryTelemetryProvider)
+    }
+    return sentryTelemetryProvider
+}
+
 private fun Variant.configureSourceBundleTasks(
     project: Project,
     extension: SentryPluginExtension,
+    sentryTelemetryProvider: Provider<SentryTelemetryService>,
     paths: OutputPaths,
     cliExecutable: String,
     sentryOrg: String?,
@@ -226,6 +276,7 @@ private fun Variant.configureSourceBundleTasks(
             val sourceContextTasks = SourceContext.register(
                 project,
                 extension,
+                sentryTelemetryProvider,
                 variant,
                 paths,
                 cliExecutable,
@@ -251,12 +302,18 @@ private fun Variant.configureSourceBundleTasks(
     }
 }
 
-private fun Variant.configureDependenciesTask(project: Project, extension: SentryPluginExtension) {
+private fun Variant.configureDependenciesTask(
+    project: Project,
+    extension: SentryPluginExtension,
+    sentryTelemetryProvider: Provider<SentryTelemetryService>
+) {
     if (isAGP74) {
         if (extension.includeDependenciesReport.get()) {
             val reportDependenciesTask =
                 SentryExternalDependenciesReportTaskFactory.register(
                     project = project,
+                    extension,
+                    sentryTelemetryProvider,
                     configurationName = "${name}RuntimeClasspath",
                     attributeValueJar = "android-classes",
                     includeReport = extension.includeDependenciesReport,
@@ -278,6 +335,7 @@ private fun Variant.configureDependenciesTask(project: Project, extension: Sentr
 private fun Variant.configureProguardMappingsTasks(
     project: Project,
     extension: SentryPluginExtension,
+    sentryTelemetryProvider: Provider<SentryTelemetryService>,
     paths: OutputPaths,
     cliExecutable: String,
     sentryOrg: String?,
@@ -293,6 +351,8 @@ private fun Variant.configureProguardMappingsTasks(
             val generateUuidTask =
                 SentryGenerateProguardUuidTask.register(
                     project = project,
+                    extension,
+                    sentryTelemetryProvider,
                     taskSuffix = name.capitalized,
                     output = paths.proguardUuidDir
                 )
@@ -300,6 +360,8 @@ private fun Variant.configureProguardMappingsTasks(
             val releaseInfo = getReleaseInfo(project, this)
             val uploadMappingsTask = SentryUploadProguardMappingsTask.register(
                 project = project,
+                extension,
+                sentryTelemetryProvider,
                 debug = extension.debug,
                 cliExecutable = cliExecutable,
                 generateUuidTask = generateUuidTask,