+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/SentryPlugin.kt
@@ -13,14 +13,13 @@ import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.plugins.ExtraPropertiesExtension
 import org.gradle.api.tasks.StopExecutionException
-import org.gradle.build.event.BuildEventsListenerRegistry
+import org.gradle.internal.build.event.BuildEventListenerRegistryInternal
 import org.slf4j.LoggerFactory
 
 @Suppress("UnstableApiUsage")
-abstract class SentryPlugin : Plugin<Project> {
-
-    @get:Inject
-    abstract val listenerRegistry: BuildEventsListenerRegistry
+abstract class SentryPlugin @Inject constructor(
+    private val buildEvents: BuildEventListenerRegistryInternal
+) : Plugin<Project> {
 
     override fun apply(project: Project) {
         if (AgpVersions.CURRENT < AgpVersions.VERSION_7_0_0) {
@@ -48,6 +47,7 @@ abstract class SentryPlugin : Plugin<Project> {
             SentryPluginExtension::class.java,
             project
         )
+
         project.pluginManager.withPlugin("com.android.application") {
             val oldAGPExtension = project.extensions.getByType(AppExtension::class.java)
             val androidComponentsExt =
@@ -68,10 +68,10 @@ abstract class SentryPlugin : Plugin<Project> {
             androidComponentsExt.configure(
                 project,
                 extension,
-                listenerRegistry,
+                buildEvents,
                 cliExecutable,
                 sentryOrgParameter,
-                sentryProjectParameter
+                sentryProjectParameter,
             )
 
             // old API configuration
@@ -80,7 +80,8 @@ abstract class SentryPlugin : Plugin<Project> {
                 extension,
                 cliExecutable,
                 sentryOrgParameter,
-                sentryProjectParameter
+                sentryProjectParameter,
+                buildEvents
             )
 
             project.installDependencies(extension, true)