+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/tasks/SentryUploadProguardMappingsTask.kt
@@ -1,6 +1,9 @@
 package io.sentry.android.gradle.tasks
 
+import io.sentry.android.gradle.extensions.SentryPluginExtension
 import io.sentry.android.gradle.tasks.SentryGenerateProguardUuidTask.Companion.SENTRY_PROGUARD_MAPPING_UUID_PROPERTY
+import io.sentry.android.gradle.telemetry.SentryTelemetryService
+import io.sentry.android.gradle.telemetry.withSentryTelemetry
 import io.sentry.android.gradle.util.PropertiesUtil
 import io.sentry.android.gradle.util.ReleaseInfo
 import io.sentry.android.gradle.util.asSentryCliExec
@@ -17,6 +20,7 @@ import org.gradle.api.tasks.Exec
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputFile
 import org.gradle.api.tasks.InputFiles
+import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.Optional
 import org.gradle.api.tasks.TaskProvider
 
@@ -64,6 +68,9 @@ abstract class SentryUploadProguardMappingsTask : Exec() {
     @get:Input
     @get:Optional
     abstract val sentryUrl: Property<String>
+
+    @get:Internal
+    abstract val sentryTelemetryService: Property<SentryTelemetryService>
 
     override fun exec() {
         if (!mappingsFiles.isPresent || mappingsFiles.get().isEmpty) {
@@ -115,6 +122,8 @@ abstract class SentryUploadProguardMappingsTask : Exec() {
         if (debug.getOrElse(false)) {
             args.add("--log-level=debug")
         }
+
+        sentryTelemetryService.orNull?.traceCli()?.let { args.addAll(it) }
 
         sentryUrl.orNull?.let {
             args.add("--url")
@@ -170,6 +179,8 @@ abstract class SentryUploadProguardMappingsTask : Exec() {
 
         fun register(
             project: Project,
+            extension: SentryPluginExtension,
+            sentryTelemetryProvider: Provider<SentryTelemetryService>?,
             debug: Property<Boolean>,
             cliExecutable: String,
             sentryProperties: String?,
@@ -202,7 +213,9 @@ abstract class SentryUploadProguardMappingsTask : Exec() {
                 task.releaseInfo.set(releaseInfo)
                 task.sentryAuthToken.set(sentryAuthToken)
                 task.sentryUrl.set(sentryUrl)
+                sentryTelemetryProvider?.let { task.sentryTelemetryService.set(it) }
                 task.asSentryCliExec()
+                task.withSentryTelemetry(extension, sentryTelemetryProvider)
             }
             return uploadSentryProguardMappingsTask
         }