+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/tasks/dependencies/SentryExternalDependenciesReportTask.kt
@@ -1,7 +1,10 @@
 package io.sentry.android.gradle.tasks.dependencies
 
+import io.sentry.android.gradle.extensions.SentryPluginExtension
 import io.sentry.android.gradle.tasks.DirectoryOutputTask
 import io.sentry.android.gradle.tasks.dependencies.SentryExternalDependenciesReportTaskFactory.SENTRY_DEPENDENCIES_REPORT_OUTPUT
+import io.sentry.android.gradle.telemetry.SentryTelemetryService
+import io.sentry.android.gradle.telemetry.withSentryTelemetry
 import io.sentry.android.gradle.util.GradleVersions
 import io.sentry.android.gradle.util.artifactsFor
 import java.io.File
@@ -77,6 +80,8 @@ abstract class SentryExternalDependenciesReportTask : DirectoryOutputTask() {
     companion object {
         fun register(
             project: Project,
+            extension: SentryPluginExtension,
+            sentryTelemetryProvider: Provider<SentryTelemetryService>,
             configurationName: String,
             attributeValueJar: String,
             output: Provider<Directory>?,
@@ -93,6 +98,7 @@ abstract class SentryExternalDependenciesReportTask : DirectoryOutputTask() {
                     project.configurations.getByName(configurationName)
                 )
                 output?.let { task.output.set(it) }
+                task.withSentryTelemetry(extension, sentryTelemetryProvider)
             }
         }
     }