+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/tasks/dependencies/SentryExternalDependenciesReportTaskV2.kt
@@ -1,7 +1,10 @@
 package io.sentry.android.gradle.tasks.dependencies
 
+import io.sentry.android.gradle.extensions.SentryPluginExtension
 import io.sentry.android.gradle.tasks.DirectoryOutputTask
 import io.sentry.android.gradle.tasks.dependencies.SentryExternalDependenciesReportTaskFactory.SENTRY_DEPENDENCIES_REPORT_OUTPUT
+import io.sentry.android.gradle.telemetry.SentryTelemetryService
+import io.sentry.android.gradle.telemetry.withSentryTelemetry
 import io.sentry.android.gradle.util.artifactsFor
 import java.io.File
 import org.gradle.api.Project
@@ -45,6 +48,8 @@ abstract class SentryExternalDependenciesReportTaskV2 : DirectoryOutputTask() {
     companion object {
         fun register(
             project: Project,
+            extension: SentryPluginExtension,
+            sentryTelemetryProvider: Provider<SentryTelemetryService>,
             configurationName: String,
             attributeValueJar: String,
             output: Provider<Directory>?,
@@ -68,6 +73,7 @@ abstract class SentryExternalDependenciesReportTaskV2 : DirectoryOutputTask() {
                 task.artifactIds.set(artifactIds)
                 task.includeReport.set(includeReport)
                 output?.let { task.output.set(it) }
+                task.withSentryTelemetry(extension, sentryTelemetryProvider)
             }
         }
     }