+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/integration/BaseSentryPluginTest.kt
@@ -104,6 +104,7 @@ abstract class BaseSentryPluginTest(
             .withArguments("--stacktrace")
             .withPluginClasspath()
             .withGradleVersion(gradleVersion)
+//            .withDebug(true)
             .forwardStdOutput(writer)
             .forwardStdError(writer)
     }